api/requirements.txt
bcrypt==3.1.7
cffi==1.13.1
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
six==1.12.0
Werkzeug==0.16.0
